Probes/kubernetes/requirements.txt
kubernetes
prometheus-client
Flask
PyQt5
grpcio
protobuf
flask-httpauth
psutil
pyyaml
